Type optimizations 32 bit is faster than 8 on ARM Unsigned is faster than signed and requires less checks

+++ pieces.h
@@ -1,4 +1,4 @@
-const uint8_t pieces[][4][16] =
+const unsigned pieces[][4][16] =
 {
 	{ // |
 		{
